more alert screen fixes
adding title and shortdescription to the top of alert template.
shortdescription is empty most of the time though.
also setting background color of alerts to transparent - it does make
fb-post (some not all) look weird, but a step in the right direction
(also the android app has the same issue so we are not alone)